Move descriptions to separate package
